cyber80_palette

+++ src/cyber80/palette/type_rgb.go
@@ -4,9 +4,9 @@ import (
 	"github.com/reiver/cyber80/src/cyber80/rgb"
 )
 
-func (receiver Type) RGB(index int) cyber80_rgb.Type {
+func (receiver Type) RGB(index byte) cyber80_rgb.Type {
 	i := index * colorSize
-	if len(receiver) <= i {
+	if len(receiver) <= int(i) {
 		return cyber80_rgb.Type{}
 	}
